move set_orientation to yaml_reader and create new GEOFlippableFileYAMLReader class
