+++ satpy/etc/readers/seviri_l1b_native.yaml
@@ -6,7 +6,7 @@ reader:
     Reader for EUMETSAT MSG SEVIRI Level 1b native format files.
   sensors: [seviri]
   default_channels: [HRV, IR_016, IR_039, IR_087, IR_097, IR_108, IR_120, IR_134, VIS006, VIS008, WV_062, WV_073]
-  reader: !!python/name:satpy.readers.yaml_reader.FileYAMLReader
+  reader: !!python/name:satpy.readers.yaml_reader.GEOFlippableFileYAMLReader
 
 file_types:
   native_msg:
